CMakeLists.txt: various fixes
Bumped minimal version of cmake to 3.5, because versions below
3.5 will be removed soon from cmake. Removed manual setting
of optimization flags. These are already set to decent defaults, and
it should be possible to set them from the outside if need be.

Also made an option SOPHUS_INSANE_CFLAGS which defaults to off. The
issue is, that (more) modern compilers will not compile the tests / code
with -Werror on. Werror is therefore harmful to production, albeit
useful for developers. Hence, it is an option to be set.

Signed-off-by: Matthias Schoepfer <matthias.schoepfer@googlemail.com>